改成nami数据

+++ app/appapi/controller/LivebackController.php
@@ -661,16 +661,18 @@ class livebackController extends HomebaseController {
                         //计算开播数量 目前仅有53个主播
                         $num = DB::name('live')->where('liveclassid', $liveClassId)->count();
                         $uid = Db::connect($sportDb)->name('sports_basketball_match_anchor')->where('match_id',$matchid)->value('user_ids');
-                        $uid_one  = Db::name('live')->where("uid={$uid}")->find();
                     }else if($liveClassId == 4){
                         //计算开播数量 目前仅有53个主播
                         $num = DB::name('live')->where('liveclassid', $liveClassId)->count();
                         $uid = Db::connect($sportDb)->name('sports_football_match_anchor')->where('match_id',$matchid)->value('user_ids');
-                        $uid_one  = Db::name('live')->where("uid={$uid}")->find();
                     }else{
                         $uid = 0;
                         $num = 0;
-                        $uid_one = 1;
+                    }
+
+                    $uid_one = 1;
+                    if($uid){
+                        $uid_one = Db::name('live')->where("uid={$uid}")->find();
                     }
 
                     if($uid && !$uid_one && $num < 100){
@@ -702,7 +704,10 @@ class livebackController extends HomebaseController {
                 }
             }
         }
-        DB::name('live')->whereIn('match_id', $closeLive)->delete();
+         if($closeLive){
+             DB::name('live')->whereIn('match_id', $closeLive)->delete();
+         }
+
     }
     
 	protected function getRandUid(){
